move list manipulating functions to `utils/Collections.kt`
